SintPowerGraphs.jl

Julia package for treating a power system as a graph.

The library is in very early development and probably only usable and useful for me. I expect the type hierarchy to undergo large changes as I learn Julia.

Method documentation

SintPowerGraphs.get_bus_data

SintPowerGraphs.get_bus_data!

SintPowerGraphs.get_id_idx

SintPowerGraphs.get_island_incidence_matrix

SintPowerGraphs.get_islanded_branches

SintPowerGraphs.get_islanded_buses

SintPowerGraphs.get_susceptance_vector

SintPowerGraphs.get_π_equivalent

SintPowerGraphs.merge_line_segments

SintPowerGraphs.remove_lines_by_value

SintPowerGraphs.subgraph

SintPowerGraphs.swapcols!

SintPowerGraphs.swaprows!

SintPowerGraphs.traverse

SintPowerCase.get_branch_data — Method

get_branch_data(network::PowerGraphBase, f_bus_id::Int, t_bus::Int)

Return a dictionary containing the dictionary with the buse data.

SintPowerCase.get_gen_indices — Method

Return indices of the buses with generators.

SintPowerCase.get_load_indices — Method

Return indices of the buses with generators.

SintPowerCase.is_gen_bus — Method

is_gen_bus(network::PowerGraphBase, bus_id::Int)

Returns true if the bus bus_id is a load.

SintPowerCase.is_load_bus — Method

is_load_bus(network::PowerGraphBase, bus_id::Int)

Returns true if the bus bus_id is a load.

SintPowerGraphs.get_bus_data! — Method

get_bus_data!(network::PowerGraphBase, bus_id::String)

Return a DataFrameRow with the bus data.

SintPowerGraphs.get_bus_data — Method

get_bus_data(network::PowerGraphBase, bus_id::Int)

Return a copy of the DataFrameRow with the bus data.

SintPowerGraphs.get_id_idx — Method

Returns the index of the element elm with index ID

SintPowerGraphs.get_island_incidence_matrix — Method

Return incidence_matrix for islands in system

SintPowerGraphs.get_islanded_branches — Method

Return branches in islands.

SintPowerGraphs.get_islanded_buses — Method

Return list of buses in islands

SintPowerGraphs.get_susceptance_vector — Method

get_susceptance_vector(network::PowerGraphBase)::Array{Float64}
Returns the susceptance vector for performing a dc power flow.

SintPowerGraphs.get_π_equivalent — Method

get_π_equivalent(network::PowerGraphBase, from_bus::Int, to_bus::Int)

Returns the π-equivalent of a line segment.

SintPowerGraphs.merge_line_segments — Method

merge_line_segments(network, keep)

	Merges consecutive line segments not separated by loads or generators.

	# Arguments
	- `network::RadialPowerGraph`: The network to merge line segments on
	- `keep_switches`::Bool=true: Keep switches.
	- `keep_indicators`::Bool=true: Keep indicators.
	- `keep_loaddata`::Bool=true: Keep loaddata.
	- `aggregators`::Dict{Symbol, Dict{Strig, Any}}=nothing: Sum up columns of the field Symbol given by String.

SintPowerGraphs.remove_lines_by_value — Function

Remove lines from a case. Whether or not a line should be removed is determined by the function func, and the tolerance tol.

SintPowerGraphs.subgraph — Function

Make a subgraph out of all descendents with a given node as root

SintPowerGraphs.swapcols! — Method

Method to efficiently swap columns taken from the internet

SintPowerGraphs.swaprows! — Method

Method to efficiently swap rows based on swapcols!

SintPowerGraphs.traverse — Function

Start at a node and traverse DFS/BFS, recording order nodes were seen Node that the order branches are traversed is not determined (the natural ordering from :args in the edges is not used...yet)